refactor: RacingCar package 위치 변경
